Redesign Entity table and Create Batch form/HTML
In entity table change the field entity_currency to currency_id.
This will be the row_id for the corresponding currency in the
currency table.

Update form to input entity with a drop down list and exclude currency
input field. During the insert of the new batch, the currency field for batch_list is inserted using a value of zero.